// src/ai/types.ts
export enum AiModelType {
    GEMINI = 'gemini',
    OLLAMA_Gemma = 'ollama-gemma',
    OLLAMA_DeepSeek = 'ollama-deepseek',
    OLLAMA_CodeLlama = 'ollama-codellama',
    OLLAMA_GPT_OSS = 'ollama-gpt-oss'
}

export enum PromptType {
    CODE_GENERATION = 'code_generation',
    GENERAL_ASK = 'general_ask'
}